+++ crowdfunding/src/main/kotlin/com/project/indistraw/crowdfunding/adapter/output/persistence/entity/AmountEntity.kt
@@ -1,14 +1,13 @@
 package com.project.indistraw.crowdfunding.adapter.output.persistence.entity
 
-import java.math.BigDecimal
 import javax.persistence.Column
 import javax.persistence.Embeddable
 
 @Embeddable
 class AmountEntity(
     @Column(nullable = false)
-    val totalAmount: BigDecimal,
+    val totalAmount: Long,
 
     @Column(nullable = false)
-    val targetAmount: BigDecimal
+    val targetAmount: Long
 )
